apps/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
PyPDF2
python-dotenv
google-generativeai
elevenlabs
sentence-transformers
httpx
